Allow Input component optional slot for label
Because I want to allow label for input to have colors or markup, I
changed the label to accept a slot. If a label prop is passed, it
will be used. If a slot is passed, it will override the label prop.

This allows all the previous label props to work as expected, but
allows optionally using styled html for the slot